refactor: separate concers and create reusable components
